Fix: pet api를 userService 및 controller에서 분리
