+++ api/src/main/java/kr/co/pawpaw/api/service/user/UserService.java
@@ -1,16 +1,9 @@
 package kr.co.pawpaw.api.service.user;
 
-import kr.co.pawpaw.api.dto.pet.CreatePetRequest;
-import kr.co.pawpaw.api.dto.pet.CreatePetResponse;
-import kr.co.pawpaw.api.dto.pet.PetResponse;
 import kr.co.pawpaw.api.dto.user.UpdateUserRequest;
 import kr.co.pawpaw.api.dto.user.UserResponse;
 import kr.co.pawpaw.api.service.file.FileService;
-import kr.co.pawpaw.common.exception.pet.NotFoundPetException;
 import kr.co.pawpaw.common.exception.user.NotFoundUserException;
-import kr.co.pawpaw.domainrdb.pet.domain.Pet;
-import kr.co.pawpaw.domainrdb.pet.service.command.PetCommand;
-import kr.co.pawpaw.domainrdb.pet.service.query.PetQuery;
 import kr.co.pawpaw.domainrdb.storage.domain.File;
 import kr.co.pawpaw.domainrdb.user.domain.User;
 import kr.co.pawpaw.domainrdb.user.domain.UserId;
@@ -20,48 +13,19 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
     private final UserQuery userQuery;
     private final FileService fileService;
-    private final PetCommand petCommand;
-    private final PetQuery petQuery;
 
     @Transactional(readOnly = true)
     public UserResponse whoAmI(final UserId userId) {
         return userQuery.findByUserId(userId)
             .map(this::getUserResponse)
             .orElseThrow(NotFoundUserException::new);
-    }
-
-    @Transactional(readOnly = true)
-    public List<PetResponse> getPetList(final UserId userId) {
-        User user = userQuery.findByUserId(userId)
-            .orElseThrow(NotFoundUserException::new);
-
-        return petQuery.findByParent(user)
-            .stream()
-            .map(PetResponse::of)
-            .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public void deletePet(
-        final UserId userId,
-        final Long petId
-    ) {
-        User user = userQuery.findByUserId(userId)
-            .orElseThrow(NotFoundUserException::new);
-
-        Pet pet = petQuery.findByParentAndId(user, petId)
-            .orElseThrow(NotFoundPetException::new);
-
-        petCommand.delete(pet);
     }
 
     @Transactional
@@ -89,17 +53,6 @@ public class UserService {
             .orElseThrow(NotFoundUserException::new);
 
         user.updateProfile(updateUserRequest.getNickname(), updateUserRequest.getBriefIntroduction());
-    }
-
-    @Transactional
-    public CreatePetResponse createPet(
-        final UserId userId,
-        final CreatePetRequest request
-    ) {
-        User user = userQuery.findByUserId(userId)
-            .orElseThrow(NotFoundUserException::new);
-
-        return CreatePetResponse.of(petCommand.save(request.toEntity(user)));
     }
 
     private UserResponse getUserResponse(final User user) {
